split current nvim config into files
